refactor(api): move route
to src/organizational-entities
