+++ api/lib/routes.js
@@ -16,7 +16,6 @@ import * as memberships from './application/memberships/index.js';
 import * as organizationInvitations from './application/organization-invitations/index.js';
 import * as organizationLearners from './application/organization-learners/index.js';
 import * as organizations from './application/organizations/index.js';
-import * as organizationsAdministration from './application/organizations-administration/index.js';
 import * as passwords from './application/passwords/index.js';
 import * as poleEmploi from './application/pole-emploi/index.js';
 import * as scoOrganizationLearners from './application/sco-organization-learners/index.js';
@@ -44,7 +43,6 @@ const routes = [
   memberships,
   organizationLearners,
   organizations,
-  organizationsAdministration,
   passwords,
   poleEmploi,
   organizationInvitations,
