+++ api/tests/organizational-entities/unit/application/organization/organization.admin.route.test.js
@@ -13,7 +13,7 @@ import {
 import { securityPreHandlers } from '../../../../../src/shared/application/security-pre-handlers.js';
 import { expect, HttpTestServer, sinon } from '../../../../test-helper.js';
 
-describe('Unit | Router | organization-router', function () {
+describe('Unit | Organizational Entities | Application | Router | organization', function () {
   let httpTestServer;
 
   beforeEach(async function () {
@@ -194,6 +194,37 @@ describe('Unit | Router | organization-router', function () {
           expect(response.statusCode).to.equal(422);
         });
       });
+    });
+  });
+
+  describe('PATCH /api/admin/organizations/{id}', function () {
+    it('returns forbidden access if admin member has CERTIF role', async function () {
+      // given
+      sinon.stub(securityPreHandlers, 'checkAdminMemberHasRoleCertif').callsFake((request, h) => h.response(true));
+      sinon
+        .stub(securityPreHandlers, 'checkAdminMemberHasRoleSuperAdmin')
+        .callsFake((request, h) => h.response({ errors: new Error('forbidden') }).code(403));
+      sinon
+        .stub(securityPreHandlers, 'checkAdminMemberHasRoleSupport')
+        .callsFake((request, h) => h.response({ errors: new Error('forbidden') }).code(403));
+      sinon
+        .stub(securityPreHandlers, 'checkAdminMemberHasRoleMetier')
+        .callsFake((request, h) => h.response({ errors: new Error('forbidden') }).code(403));
+
+      // when
+      const response = await httpTestServer.request('PATCH', '/api/admin/organizations/1', {
+        data: {
+          id: '1',
+          type: 'organizations',
+          attributes: {
+            name: 'Super Tag',
+            type: 'SCO',
+          },
+        },
+      });
+
+      // then
+      expect(response.statusCode).to.equal(403);
     });
   });
 });
