--- api/tests/unit/application/organizations-administration/index_test.js
@@ -1,39 +1,0 @@
-import * as moduleUnderTest from '../../../../lib/application/organizations-administration/index.js';
-import { securityPreHandlers } from '../../../../src/shared/application/security-pre-handlers.js';
-import { expect, HttpTestServer, sinon } from '../../../test-helper.js';
-
-describe('Unit | Router | organization-administration-router', function () {
-  describe('PATCH /api/admin/organizations/{id}', function () {
-    it('returns forbidden access if admin member has CERTIF role', async function () {
-      // given
-      sinon.stub(securityPreHandlers, 'checkAdminMemberHasRoleCertif').callsFake((request, h) => h.response(true));
-      sinon
-        .stub(securityPreHandlers, 'checkAdminMemberHasRoleSuperAdmin')
-        .callsFake((request, h) => h.response({ errors: new Error('forbidden') }).code(403));
-      sinon
-        .stub(securityPreHandlers, 'checkAdminMemberHasRoleSupport')
-        .callsFake((request, h) => h.response({ errors: new Error('forbidden') }).code(403));
-      sinon
-        .stub(securityPreHandlers, 'checkAdminMemberHasRoleMetier')
-        .callsFake((request, h) => h.response({ errors: new Error('forbidden') }).code(403));
-
-      const httpTestServer = new HttpTestServer();
-      await httpTestServer.register(moduleUnderTest);
-
-      // when
-      const response = await httpTestServer.request('PATCH', '/api/admin/organizations/1', {
-        data: {
-          id: '1',
-          type: 'organizations',
-          attributes: {
-            name: 'Super Tag',
-            type: 'SCO',
-          },
-        },
-      });
-
-      // then
-      expect(response.statusCode).to.equal(403);
-    });
-  });
-});
